Update the displayController module for the game
Updates the displayController module which displays the game and game
board information onto the page. Also makes some minor changes in
existing functions in the gameController factory function and introduces
small naming changes for select variables throughout the script. As for
the css and html, a few headers and some simple styling are added.