Add the functionality to change player names
Adds player info to the display for the game and provides the
functionality to change player names. Also introduces new html elements
to accomodate all of the additional info to be displayed along with
additional styling for a new layout. Some minor changes and additions to
logic for the name change to function correctly are also introduced.